# contracts/openapi-v1.baseline.json
{
  "components": {
    "schemas": {
...
      "RejectApplicationRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "RequestInfoRequest": {
        "additionalProperties": false,
        "properties": {
          "flaggedDocumentTypeCodes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "flaggedProfileFields": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason",
          "flaggedProfileFields",
          "flaggedDocumentTypeCodes"
        ],
        "type": "object"
      },
...
      "SupplierLifecycleRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MotsSupplierPortal.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/review/queue": {
      "get": {
        "description": "Requires the `supplier.review` permission.",
        "operationId": "ListReviewQueue",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "withCount",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "assignedTo",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.review"
        ]
      }
    },
...
    "/api/v1/review/{referenceCode}": {
      "get": {
        "description": "Requires the `supplier.review` permission.",
        "operationId": "GetReviewerSupplierView",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "The version of this resource, to be sent back as If-Match on a write.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.review"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/approve": {
      "post": {
        "description": "Requires the `supplier.approve` permission.",
        "operationId": "ApproveApplication",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The ETag of the version being edited, from a prior read of this resource or of the aggregate containing it. Missing: 428. Not a version this API issued, or no longer current: 412.",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "The version of this resource, to be sent back as If-Match on a write.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN"
          },
          "412": {
            "description": "ETAG_MISMATCH"
          },
          "428": {
            "description": "IF_MATCH_REQUIRED"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.approve"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/claim": {
      "post": {
        "description": "Requires the `supplier.review` permission.",
        "operationId": "ClaimReviewItem",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.review"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/deactivate": {
      "post": {
        "description": "Requires the `supplier.lifecycle.manage` permission.",
        "operationId": "DeactivateSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierLifecycleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.lifecycle.manage"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/pickup": {
      "post": {
        "description": "Requires the `supplier.review` permission.",
        "operationId": "PickUpApplication",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.review"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/reactivate": {
      "post": {
        "description": "Requires the `supplier.lifecycle.manage` permission.",
        "operationId": "ReactivateSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierLifecycleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.lifecycle.manage"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/reject": {
      "post": {
        "description": "Requires the `supplier.reject` permission.",
        "operationId": "RejectApplication",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The ETag of the version being edited, from a prior read of this resource or of the aggregate containing it. Missing: 428. Not a version this API issued, or no longer current: 412.",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "The version of this resource, to be sent back as If-Match on a write.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN"
          },
          "412": {
            "description": "ETAG_MISMATCH"
          },
          "428": {
            "description": "IF_MATCH_REQUIRED"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.reject"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/request-info": {
      "post": {
        "description": "Requires the `supplier.requestInfo` permission.",
        "operationId": "RequestApplicationInfo",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The ETag of the version being edited, from a prior read of this resource or of the aggregate containing it. Missing: 428. Not a version this API issued, or no longer current: 412.",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestInfoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "The version of this resource, to be sent back as If-Match on a write.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN"
          },
          "412": {
            "description": "ETAG_MISMATCH"
          },
          "428": {
            "description": "IF_MATCH_REQUIRED"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.requestInfo"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/retry-erp-push": {
      "post": {
        "description": "Requires the `admin.integrations.manage` permission.",
        "operationId": "RetryErpSupplierPush",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "admin.integrations.manage"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/suspend": {
      "post": {
        "description": "Requires the `supplier.lifecycle.manage` permission.",
        "operationId": "SuspendSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierLifecycleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.lifecycle.manage"
        ]
      }
    },
    "/api/v1/review/{referenceCode}/unassign": {
      "post": {
        "description": "Requires the `supplier.review` permission.",
        "operationId": "UnassignReviewItem",
        "parameters": [
          {
            "in": "path",
            "name": "referenceCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "tags": [
          "Review"
        ],
        "x-required-permissions": [
          "supplier.review"
        ]
      }
    },
...
    "/api/v1/suppliers/me/resubmit-application": {
      "post": {
        "description": "Requires the `supplier.edit` permission.",
        "operationId": "ResubmitApplication",
        "parameters": [
          {
            "description": "The ETag of the version being edited, from a prior read of this resource or of the aggregate containing it. Missing: 428. Not a version this API issued, or no longer current: 412.",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "The version of this resource, to be sent back as If-Match on a write.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN"
          },
          "412": {
            "description": "ETAG_MISMATCH"
          },
          "428": {
            "description": "IF_MATCH_REQUIRED"
          }
        },
        "tags": [
          "Suppliers"
        ],
        "x-required-permissions": [
          "supplier.edit"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}